refactor(formatter): split generated ownership boundary
Keep the generator-facing formatter API in formatter.rs and move the hand-maintained rendering implementation into formatter_ext.rs. Repoint rate-limit formatting at the extension module so future code generation cannot overwrite these helpers.
